+++ cherab/generomak/plasma/plasma.py
@@ -1,7 +1,7 @@
 
-# Copyright 2016-2021 Euratom
-# Copyright 2016-2021 United Kingdom Atomic Energy Authority
-# Copyright 2016-2021 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
 #
 # Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
 # European Commission - subsequent versions of the EUPL (the "Licence");
@@ -24,7 +24,8 @@ from scipy.constants import atomic_mass, electron_mass
 
 from raysect.core import Vector3D, translate
 from raysect.core.math.function.float.function2d.interpolate import Discrete2DMesh
-from raysect.core.math.function.float import Arg1D, Exp1D, Constant1D
+from raysect.core.math.function.float import Arg1D, Exp1D, Constant1D, Interpolator1DArray, Blend2D
+from raysect.core.math.function.vector3d import Constant2D as ConstantVector2D, Blend2D as BlendVector2D
 from raysect.primitive import Cylinder, Subtract
 
 from cherab.core import AtomicData, Plasma, Maxwellian, Species
@@ -106,64 +107,66 @@ def get_edge_interpolators():
                         profiles["mesh"]["triangles"],
                         profiles["electron"]["temperature"], limit=False)
     ne = Discrete2DMesh.instance(te, profiles["electron"]["density"], limit=False)
+    ve = ConstantVector2D(Vector3D(0, 1.e-10, 0))  # avoid zero-length vectors for blending
 
     mesh_interp["electron"]["temperature"] = te
     mesh_interp["electron"]["density"] = ne
+    mesh_interp["electron"]["velocity"] = ve
 
     for elem_name, elem_data in profiles["composition"].items():
         for stage, stage_data in elem_data.items():
 
             t = Discrete2DMesh.instance(te, stage_data["temperature"], limit=False)
             n = Discrete2DMesh.instance(te, stage_data["density"], limit=False)
+            v = ConstantVector2D(Vector3D(0, 1.e-10, 0))  # avoid zero-length vectors for blending
 
             mesh_interp["composition"][elem_name][stage]["temperature"] = t
             mesh_interp["composition"][elem_name][stage]["density"] = n
+            mesh_interp["composition"][elem_name][stage]["velocity"] = v
             mesh_interp["composition"][elem_name][stage]["element"] = stage_data["element"]
 
     return mesh_interp.freeze()
 
 
-def get_edge_distributions():
+def get_2d_distributions(profiles_2d=None):
     """
-    Provides Generomak edge Maxwellian distribution of plasma species
+    Provides Generomak Maxwellian distribution of plasma species for 2d profiles
 
+    :param profiles_2d: Dictionary with 2D profile interpolators in the shape
+                        returned by the get_edge_interpolators() or get_full_profiles() functions.
+                        If not specified, will use the value returned by get_edge_interpolators().
     :return: Dictionary holding instances of Maxwellian distributions for plasma species.
     """
 
-    mesh_interp = get_edge_interpolators()
-
-    zero_vector = Vector3D(0, 0, 0)
+    profiles_2d = profiles_2d or get_edge_interpolators()
 
     dists = RecursiveDict()
 
-    n3d = AxisymmetricMapper(mesh_interp["electron"]["density"])
-    t3d = AxisymmetricMapper(mesh_interp["electron"]["temperature"])
+    n3d = AxisymmetricMapper(profiles_2d["electron"]["density"])
+    t3d = AxisymmetricMapper(profiles_2d["electron"]["temperature"])
+    v3d = VectorAxisymmetricMapper(profiles_2d["electron"]["velocity"])
 
-    dists["electron"] = Maxwellian(n3d, t3d, zero_vector, electron_mass)
+    dists["electron"] = Maxwellian(n3d, t3d, v3d, electron_mass)
 
-    for elem_name, elem_data in mesh_interp["composition"].items():
+    for elem_name, elem_data in profiles_2d["composition"].items():
         for stage, stage_data in elem_data.items():
 
-            # get element or isotope
-            try:
-                element = lookup_isotope(elem_name)
-            except ValueError:
-                element = lookup_element(elem_name)
+            spec_cherab = _get_cherab_element(elem_name)
 
             n3d = AxisymmetricMapper(stage_data["density"])
             t3d = AxisymmetricMapper(stage_data["temperature"])
-            mass = element.atomic_weight * atomic_mass
-            dists["composition"][elem_name][stage]["distribution"] = Maxwellian(n3d, t3d, zero_vector, mass)
-            dists["composition"][elem_name][stage]["element"] = element
+            v3d = VectorAxisymmetricMapper(stage_data["velocity"])
+            mass = spec_cherab.atomic_weight * atomic_mass
+            dists["composition"][elem_name][stage] = Maxwellian(n3d, t3d, v3d, mass)
 
     return dists.freeze()
 
 
 def get_edge_plasma(atomic_data=None, parent=None, name="Generomak edge plasma"):
     """
-    Provides Generomak Edge plasma.
+    Provides Generomak default edge plasma.
 
-    :param atomic_data: Instance of AtomicData, default isOpenADAS()
+    :param atomic_data: Instance of AtomicData, default is OpenADAS()
     :param parent: parent of the plasma node, defaults None
     :param name: name of the plasma node, defaults "Generomak edge plasma"
     :return: populated Plasma object
@@ -172,12 +175,8 @@ def get_edge_plasma(atomic_data=None, parent=None, name="Generomak edge plasma")
     # load Generomak equilibrium
     equilibrium = load_equilibrium()
 
-    # create or check atomic_data
-    if atomic_data is not None:
-        if not isinstance(atomic_data, AtomicData):
-            raise ValueError("atomic_data has to be of type AtomicData")
-    else:
-        atomic_data = OpenADAS()
+    # get edge distributions
+    distributions = get_2d_distributions()
 
     # base plasma geometry on mesh vertices
     profiles_dir = os.path.join(os.path.dirname(__file__), "data/edge")
@@ -188,38 +187,10 @@ def get_edge_plasma(atomic_data=None, parent=None, name="Generomak edge plasma")
     vertex_coords = np.asarray(mesh["vertex_coords"])
     r_range = (vertex_coords[:, 0].min(), vertex_coords[:, 0].max())
     z_range = (vertex_coords[:, 1].min(), vertex_coords[:, 1].max())
-    plasma_height = z_range[1] - z_range[0]
 
-    padding = 1e-3  # enlarge for safety
-
-    outer_column = Cylinder(radius=r_range[1], height=plasma_height)
-    inner_column = Cylinder(radius=r_range[0], height=plasma_height + 2 * padding)
-    inner_column.transform = translate(0, 0, -padding)
-
-    plasma_geometry = Subtract(outer_column, inner_column)
-    geometry_transform = translate(0, 0, z_range[0])
-
-    # get distributions
-    dists = get_edge_distributions()
-
-    # create plasma composition list
-    plasma_composition = []
-    for elem_data in dists["composition"].values():
-        for stage, stage_data in elem_data.items():
-            species = Species(stage_data["element"], stage, stage_data["distribution"])
-            plasma_composition.append(species)
-
-    # Populate plasma
-    plasma = Plasma(parent=parent)
-    plasma.name = name
-    plasma.geometry = plasma_geometry
-    plasma.atomic_data = atomic_data
-    plasma.electron_distribution = dists["electron"]
-    plasma.composition = plasma_composition
-    plasma.geometry_transform = geometry_transform
-    plasma.b_field = VectorAxisymmetricMapper(equilibrium.b_field)
-
-    return plasma
+    return get_plasma(equilibrium=equilibrium, distributions=distributions,
+                      r_range=r_range, z_range=z_range, atomic_data=atomic_data,
+                      parent=parent, name=name)
 
 
 def get_double_parabola(v_min, v_max, convexity, concavity, xmin=0, xmax=1):
@@ -245,7 +216,7 @@ def get_double_parabola(v_min, v_max, convexity, concavity, xmin=0, xmax=1):
     :return: Function1D
     """
 
-    x = Arg1D() #the free parameter
+    x = Arg1D()  # the free parameter
 
     # funciton for the normalised free variable
     x_norm = ClampInput1D((x - xmin) / (xmax - xmin), 0, 1)
@@ -271,7 +242,7 @@ def get_exponential_growth(initial_value, growth_rate, initial_position=1):
     :return: Function1D
     """
 
-    x = Arg1D() #the free parameter
+    x = Arg1D()  # the free parameter
     return initial_value * Exp1D((x - initial_position) * growth_rate)
 
 
@@ -287,7 +258,6 @@ def get_maxwellian_distribution(equilibrium, f1d_density, f1d_temperature, f1d_v
     :rest_mass: Rest mass of the distribution species.
     :return: Maxwellian distribution
     """
-
 
     # map profiles to 3D
     f3d_te = equilibrium.map3d(f1d_temperature)
@@ -308,28 +278,29 @@ def get_edge_profile_values(r, z, edge_interpolators=None):
            returned by the get_edge_interpolators function.
     :return: Dictionary of edge values at [R, Z]
     """
+
     # load edge interpolators if not passed as argument
     if edge_interpolators is None:
         edge_interp = get_edge_interpolators()
     else:
         edge_interp = edge_interpolators
     # create recursive dictionary to store profile values
-    lcfs_values = RecursiveDict()
+    values = RecursiveDict()
 
     # add electron values
-    lcfs_values["electron"]["temperature"] = edge_interp["electron"]["temperature"](r, z)
-    lcfs_values["electron"]["density"] = edge_interp["electron"]["density"](r, z)
+    values["electron"]["temperature"] = edge_interp["electron"]["temperature"](r, z)
+    values["electron"]["density"] = edge_interp["electron"]["density"](r, z)
 
     # add species values
     for spec, desc in edge_interp['composition'].items():
         for chrg, chrg_desc in desc.items():
             for prop, val in chrg_desc.items():
                 if prop in ["temperature", "density"]:
-                    lcfs_values["composition"][spec][chrg][prop] = val(r, z)
+                    values["composition"][spec][chrg][prop] = val(r, z)
                 else:
-                    lcfs_values["composition"][spec][chrg][prop] = val
+                    values["composition"][spec][chrg][prop] = val
 
-    return lcfs_values.freeze()
+    return values.freeze()
 
 
 def get_core_profiles_arguments(**kwargs):
@@ -371,26 +342,26 @@ def get_core_profiles_arguments(**kwargs):
 
     :return: dictionary of profile arguments
     """
-    
+
     core_args = {"ne_core": 5e19, "ne_convexity": 1.09,
-                         "ne_concavity": 0.24, "te_core": 3e3,
-                         "te_convexity": 2.35, "te_concavity": 1.26,
-                         "nh_core": 5e19, "nh_convexity": 1.09,
-                         "nh_concavity": 0.24, "th_core": 2.8e3,
-                         "th_convexity": 1, "th_concavity": 0.82,
-                         "th0_fraction": 0.8, "nh0_decay": 20,
-                         "timp_core": 2.7e3, "timp_convexity": 1,
-                         "timp_concavity": 0.82, "nimp_core": 5e17,
-                         "nimp_convexity": 1.09, "nimp_concavity": 0.24,
-                         "nimp_decay": 30,
-                         "vtor_core": 1e5, "vtor_edge": 1e4,
-                         "vtor_convexity": 2, "vtor_concavity": 4,
-                         "vpol_lcfs": 2e4, "vpol_decay": 0.08}
+                 "ne_concavity": 0.24, "te_core": 3e3,
+                 "te_convexity": 2.35, "te_concavity": 1.26,
+                 "nh_core": 5e19, "nh_convexity": 1.09,
+                 "nh_concavity": 0.24, "th_core": 2.8e3,
+                 "th_convexity": 1, "th_concavity": 0.82,
+                 "th0_fraction": 0.8, "nh0_decay": 20,
+                 "timp_core": 2.7e3, "timp_convexity": 1,
+                 "timp_concavity": 0.82, "nimp_core": 5e17,
+                 "nimp_convexity": 1.09, "nimp_concavity": 0.24,
+                 "nimp_decay": 30,
+                 "vtor_core": 1e5, "vtor_edge": 1e4,
+                 "vtor_convexity": 2, "vtor_concavity": 4,
+                 "vpol_lcfs": 2e4, "vpol_decay": 0.08}
 
     if not kwargs:
         return core_args
 
-   # change passed values of core args
+    # change passed values of core args
     for key, item in kwargs.items():
         core_args[key] = item
 
@@ -419,15 +390,23 @@ def get_core_profiles_description(lcfs_values=None, core_args=None):
         z = 0
         lcfs_values = get_edge_profile_values(r, z)
 
+        # manually correcting the LCFS densities for neutral species and low ionised carbon
+        # to better match the edge profiles
+        lcfs_values["composition"]["hydrogen"][0]["density"] = 1.e15
+        lcfs_values["composition"]["carbon"][0]["density"] = 1.e5
+        lcfs_values["composition"]["carbon"][1]["density"] = 1.5e5
+        lcfs_values["composition"]["carbon"][2]["density"] = 1.e9
+        lcfs_values["composition"]["carbon"][3]["density"] = 1.e13
+
     if core_args is None:
         core_args = get_core_profiles_arguments()
 
     # toroidal rotation profile
     f1d_vtor = get_double_parabola(core_args["vtor_edge"], core_args["vtor_core"],
-                                   core_args["vtor_convexity"], core_args["vtor_concavity"])
+                                   core_args["vtor_convexity"], core_args["vtor_concavity"], xmin=1, xmax=0)
 
     # poloidal rotation profile
-    f1d_vpol = get_exponential_growth(core_args["vpol_lcfs"], core_args["vpol_decay"]) 
+    f1d_vpol = get_exponential_growth(core_args["vpol_lcfs"], core_args["vpol_decay"])
 
     # velocity normal to magnetic surfaces
     f1d_vnorm = Constant1D(0)
@@ -442,7 +421,7 @@ def get_core_profiles_description(lcfs_values=None, core_args=None):
     profiles["electron"]["f1d_density"] = get_double_parabola(lcfs_values["electron"]["density"],
                                                               core_args["ne_core"], core_args["ne_convexity"],
                                                               core_args["ne_concavity"], xmin=1, xmax=0)
-    profiles["electron"]["f1d_vtor"] = Constant1D(0)
+    profiles["electron"]["f1d_vtor"] = Constant1D(1.e-10)  # avoid zero-length vectors for blending
     profiles["electron"]["f1d_vpol"] = Constant1D(0)
     profiles["electron"]["f1d_vnorm"] = Constant1D(0)
 
@@ -461,7 +440,8 @@ def get_core_profiles_description(lcfs_values=None, core_args=None):
     profiles["composition"]["hydrogen"][0]["f1d_temperature"] = core_args["th0_fraction"] * get_double_parabola(lcfs_values["composition"]["hydrogen"][0]["temperature"],
                                                                                                                 core_args["th_core"], core_args["th_convexity"],
                                                                                                                 core_args["th_concavity"], xmin=1, xmax=0)
-    profiles["composition"]["hydrogen"][0]["f1d_density"] = get_exponential_growth(lcfs_values["composition"]["hydrogen"][0]["density"], core_args["nh0_decay"])
+    profiles["composition"]["hydrogen"][0]["f1d_density"] = get_exponential_growth(
+        lcfs_values["composition"]["hydrogen"][0]["density"], core_args["nh0_decay"])
     profiles["composition"]["hydrogen"][0]["f1d_vtor"] = f1d_vtor
     profiles["composition"]["hydrogen"][0]["f1d_vpol"] = f1d_vpol
     profiles["composition"]["hydrogen"][0]["f1d_vnorm"] = f1d_vnorm
@@ -471,7 +451,7 @@ def get_core_profiles_description(lcfs_values=None, core_args=None):
                                                                                   core_args["timp_core"], core_args["timp_convexity"], core_args["timp_concavity"],
                                                                                   xmin=1, xmax=0)
     profiles["composition"]["carbon"][6]["f1d_density"] = get_double_parabola(lcfs_values["composition"]["carbon"][6]["density"], core_args["nimp_core"],
-                                                                              core_args["nimp_convexity"], core_args["nimp_concavity"], xmin=1, xmax=0) 
+                                                                              core_args["nimp_convexity"], core_args["nimp_concavity"], xmin=1, xmax=0)
     profiles["composition"]["carbon"][6]["f1d_vtor"] = f1d_vtor
     profiles["composition"]["carbon"][6]["f1d_vpol"] = f1d_vpol
     profiles["composition"]["carbon"][6]["f1d_vnorm"] = f1d_vnorm
@@ -481,7 +461,8 @@ def get_core_profiles_description(lcfs_values=None, core_args=None):
         profiles["composition"]["carbon"][chrg]["f1d_temperature"] = get_double_parabola(lcfs_values["composition"]["carbon"][chrg]["temperature"],
                                                                                          core_args["timp_core"], core_args["timp_convexity"], core_args["timp_concavity"],
                                                                                          xmin=1, xmax=0)
-        profiles["composition"]["carbon"][chrg]["f1d_density"] = get_exponential_growth(lcfs_values["composition"]["carbon"][chrg]["density"], core_args["nimp_decay"])
+        profiles["composition"]["carbon"][chrg]["f1d_density"] = get_exponential_growth(
+            lcfs_values["composition"]["carbon"][chrg]["density"], core_args["nimp_decay"])
         profiles["composition"]["carbon"][chrg]["f1d_vtor"] = f1d_vtor
         profiles["composition"]["carbon"][chrg]["f1d_vpol"] = f1d_vpol
         profiles["composition"]["carbon"][chrg]["f1d_vnorm"] = f1d_vnorm
@@ -493,8 +474,8 @@ def get_core_distributions(profiles=None, equilibrium=None):
     """
     Returns a dictionary of core plasma species Maxwellian distributions.
 
-    :param profiles: Dictionary of core particle profiles.  The dictionary has to have the same form 
-                     as the one returned by the function get_core_profiles_description. 
+    :param profiles: Dictionary of core particle profiles.  The dictionary has to have the same form
+                     as the one returned by the function get_core_profiles_description.
                      The default value is the value returned by the call get_core_profiles_description().
     :param equilibrium: an instance of EFITEquilibrium.
     :return:  dictionary of core plasma species with Maxwellian distribution
@@ -510,23 +491,20 @@ def get_core_distributions(profiles=None, equilibrium=None):
     # build a dictionary with Maxwellian distributions
     species = RecursiveDict()
     species["electron"] = get_maxwellian_distribution(equilibrium, rest_mass=electron_mass,
-                                                            **profiles["electron"])
+                                                      **profiles["electron"])
     for name, spec in profiles["composition"].items():
         spec_cherab = _get_cherab_element(name)
         for chrg, desc in spec.items():
             rest_mass = atomic_mass * spec_cherab.atomic_weight
-            species["composition"][name][chrg] = get_maxwellian_distribution(equilibrium, rest_mass=rest_mass,  **desc)
+            species["composition"][name][chrg] = get_maxwellian_distribution(equilibrium, rest_mass=rest_mass, **desc)
 
     return species.freeze()
 
 
-def get_core_plasma(distributions=None, atomic_data=None, parent=None, name="Generomak core plasma"):
+def get_core_plasma(atomic_data=None, parent=None, name="Generomak core plasma"):
     """
-    Provides Generomak core plasma.
+    Provides Generomak default core plasma.
 
-    :param distributions: A dictionary of plasma distributions. Has to have the same format as the
-                          dictionary returned by get_core_distributions. The default value
-                          is the value returned by the call get_core_distributions().
     :param atomic_data: Instance of AtomicData, default is OpenADAS()
     :param parent: parent of the plasma node, defaults None
     :param name: name of the plasma node, defaults "Generomak edge plasma"
@@ -536,33 +514,119 @@ def get_core_plasma(distributions=None, atomic_data=None, parent=None, name="Gen
     # load Generomak equilibrium
     equilibrium = load_equilibrium()
 
+    # load core distributions
+    distributions = get_core_distributions(equilibrium=equilibrium)
+
+    return get_plasma(equilibrium=equilibrium, distributions=distributions,
+                      atomic_data=atomic_data, parent=parent, name=name)
+
+
+def get_full_profiles(equilibrium=None, core_profiles=None, edge_profiles=None, mask=None):
+    """
+    Blends core and edge profiles using the mask function as a modulator.
+
+    :param equilibrium: an instance of EFITEquilibrium. The default value is the value returned by
+                        load_equilibrium().
+    :param core_profiles: Dictionary of core particle profiles.  The dictionary has to have
+                          the same form as the one returned by the function
+                          get_core_profiles_description.
+                          The default value is the value returned by the call
+                          get_core_profiles_description().
+    :param edge_profiles: Dictionary with edge interpolators in the shape
+                          returned by the get_edge_interpolators function.
+                          If not specified, will use the value returned by
+                          get_edge_interpolators().
+    :param Function2D mask: Scalar 2D function returning a value in the range [0, 1].
+                            If not specified, will use core profiles for psi_normal < 0.94,
+                            the edge profiles for psi_normal > 1 and a weighted sum of core and
+                            edge profiles for 0.94 < psi_normal < 1, with the edge profile weight
+                            increasing from 0 to 1 linearly.
+
+    :return: dictionary of blended plasma profiles with the sturcture identical to edge_profiles.
+    """
+
+    equilibrium = equilibrium or load_equilibrium()
+
+    core_profiles = core_profiles or get_core_profiles_description()
+
+    edge_profiles = edge_profiles or get_edge_interpolators()
+
+    mask = mask or equilibrium.map2d(Interpolator1DArray([0, 0.94, 1.0, 1.1], [1, 1, 0, 0], 'linear', 'none', 0))
+
+    # blended core and edge profiles
+    blended_profiles = RecursiveDict()
+
+    # map core profiles to 2D using the equilibrium
+    te_core = equilibrium.map2d(core_profiles["electron"]["f1d_temperature"])
+    ne_core = equilibrium.map2d(core_profiles["electron"]["f1d_density"])
+    ve_core = equilibrium.map_vector2d(core_profiles["electron"]["f1d_vtor"],
+                                       core_profiles["electron"]["f1d_vpol"],
+                                       core_profiles["electron"]["f1d_vnorm"])
+
+    blended_profiles["electron"]["temperature"] = Blend2D(edge_profiles["electron"]["temperature"], te_core, mask)
+    blended_profiles["electron"]["density"] = Blend2D(edge_profiles["electron"]["density"], ne_core, mask)
+    blended_profiles["electron"]["velocity"] = BlendVector2D(edge_profiles["electron"]["velocity"], ve_core, mask)
+
+    for element, states in core_profiles["composition"].items():
+        for charge, state in states.items():
+            t_core = equilibrium.map2d(state["f1d_temperature"])
+            n_core = equilibrium.map2d(state["f1d_density"])
+            v_core = equilibrium.map_vector2d(state["f1d_vtor"], state["f1d_vpol"], state["f1d_vnorm"])
+
+            edge_state = edge_profiles["composition"][element][charge]
+
+            blended_profiles["composition"][element][charge]["temperature"] = Blend2D(edge_state["temperature"], t_core, mask)
+            blended_profiles["composition"][element][charge]["density"] = Blend2D(edge_state["density"], n_core, mask)
+            blended_profiles["composition"][element][charge]["velocity"] = BlendVector2D(edge_state["velocity"], v_core, mask)
+
+    return blended_profiles.freeze()
+
+
+def get_plasma(equilibrium=None, distributions=None, r_range=None, z_range=None, atomic_data=None, parent=None, name="Generomak plasma"):
+    """
+    Provides Generomak plasma. The full (core + edge) plasma is returned by default.
+
+    :param equilibrium: an instance of EFITEquilibrium. The default value is the value returned by load_equilibrium().
+    :param distributions: A dictionary of plasma distributions. Has to have the same format as the
+                          dictionary returned by get_core_distributions or get_2d_distributions.
+                          The default value is the value returned by the call:
+                          get_2d_distributions(get_full_profiles(equilibrium)).
+    :param r_range: Plasma domain range (min, max) in R direction in meters.
+    :param z_range: Plasma domain range (min, max) in Z direction in meters.
+    :param atomic_data: Instance of AtomicData, default is OpenADAS()
+    :param parent: parent of the plasma node, defaults None
+    :param name: name of the plasma node, defaults "Generomak plasma"
+    :return: populated Plasma object
+    """
+
+    equilibrium = equilibrium or load_equilibrium()
+
+    distributions = distributions or get_2d_distributions(get_full_profiles(equilibrium=equilibrium))
+
+    r_range = r_range or equilibrium.r_range
+    z_range = z_range or equilibrium.z_range
+
     # create or check atomic_data
     if atomic_data is not None:
         if not isinstance(atomic_data, AtomicData):
-            raise ValueError("atomic_data has to be of type AtomicData")   
+            raise ValueError("atomic_data has to be of type AtomicData")
     else:
         atomic_data = OpenADAS()
 
-    # construct plasma primitive shape 
-    padding = 1e-3 #enlarge for safety
-    plasma_height = equilibrium.z_range[1] - equilibrium.z_range[0]
-    outer_column = Cylinder(radius=equilibrium.r_range[1], height=plasma_height)
-    inner_column = Cylinder(radius=equilibrium.r_range[0], height=plasma_height + 2 * padding)
+    # construct plasma primitive shape
+    padding = 1e-3  # enlarge for safety
+    plasma_height = z_range[1] - z_range[0]
+    outer_column = Cylinder(radius=r_range[1], height=plasma_height)
+    inner_column = Cylinder(radius=r_range[0], height=plasma_height + 2 * padding)
     inner_column.transform = translate(0, 0, -padding)
     plasma_geometry = Subtract(outer_column, inner_column)
 
     # coordinate transform of the plasma frame
-    geometry_transform = translate(0, 0, equilibrium.z_range[0])
-
-    # load core distributions if needed
-    if distributions is None:
-        dists = get_core_distributions()
-    else:
-        dists = distributions
+    geometry_transform = translate(0, 0, z_range[0])
 
     # create plasma composition list
     plasma_composition = []
-    for elem_name, elem_data in dists["composition"].items():
+    for elem_name, elem_data in distributions["composition"].items():
         for stage, stage_data in elem_data.items():
             elem = _get_cherab_element(elem_name)
             species = Species(elem, stage, stage_data)
@@ -573,7 +637,7 @@ def get_core_plasma(distributions=None, atomic_data=None, parent=None, name="Gen
     plasma.name = name
     plasma.geometry = plasma_geometry
     plasma.atomic_data = atomic_data
-    plasma.electron_distribution = dists["electron"]
+    plasma.electron_distribution = distributions["electron"]
     plasma.composition = plasma_composition
     plasma.geometry_transform = geometry_transform
     plasma.b_field = VectorAxisymmetricMapper(equilibrium.b_field)